+++ etc/uikit-playground/src/test/index.vue
@@ -2,7 +2,7 @@
 import { ref } from 'vue';
 import Layout from '@/Layout.vue';
 import Scrollable from '@/test/Scrollable.vue';
-import { PlTooltip, useTheme, useResizeObserver, MaskIcon16 } from '@milaboratories/uikit';
+import { PlTooltip, useTheme, useResizeObserver, PlMaskIcon16 } from '@milaboratories/uikit';
 
 const resizable = ref<HTMLElement | undefined>();
 
@@ -26,8 +26,8 @@ function onCloseTooltip() {
       <div>Item 2</div>
       <div>Item 3</div>
       <div>
-        <MaskIcon16 style="background-color: red" name="play" />
-        <MaskIcon16 style="background-color: red" name="checkmark" />
+        <PlMaskIcon16 style="background-color: red" name="play" />
+        <PlMaskIcon16 style="background-color: red" name="checkmark" />
       </div>
     </div>
     <div style="display: flex; background-color: #fff; padding: 12px; margin-top: 12px">