+++ lib/ui/uikit/src/components/PlIcon16/index.ts
@@ -1,0 +1,1 @@
+export { default as PlIcon16 } from './PlIcon16.vue';