+++ lib/ui/uikit/src/components/PlLogView/PlLogView.vue
@@ -8,12 +8,14 @@ export default {
 </script>
 
 <script lang="ts" setup>
-import { computed, onMounted, onUpdated, ref } from 'vue';
-import MaskIcon24 from '../MaskIcon24.vue';
+import { computed, ref, watch } from 'vue';
+import { PlMaskIcon24 } from '../PlMaskIcon24';
 import './pl-log-view.scss';
 import { okOptional, tapIf } from '@milaboratories/helpers';
 import type { AnyLogHandle, Platforma, ValueOrErrors } from '@platforma-sdk/model';
 import { useLogHandle } from './useLogHandle';
+import { useLabelNotch } from '@/utils/useLabelNotch';
+import DoubleContour from '@/utils/DoubleContour.vue';
 
 const getOutputError = <T,>(o?: ValueOrErrors<T>) => {
   if (o && o.ok === false) {
@@ -43,18 +45,32 @@ const props = defineProps<{
    */
   output?: ValueOrErrors<unknown>;
   /**
+   * Max retries for AnyLogHandle fetch (with the same parameters)
+   */
+  maxRetries?: number;
+  /**
    * @TODO
    */
   mockPlatforma?: Platforma;
+  /**
+   * The label to display above the texarea.
+   */
+  label?: string;
 }>();
 
 const logState = useLogHandle(props);
 
+const isAnchored = ref<boolean>(true);
+
 const contentRef = ref<HTMLElement>();
+
+const root = ref<HTMLInputElement>();
 
 const computedError = computed(() => logState.value?.error ?? props.error ?? getOutputError(props.output));
 
 const computedValue = computed(() => logState.value?.lines ?? props.value ?? okOptional(props.output));
+
+useLabelNotch(root);
 
 const onClickCopy = () => {
   if (computedValue.value && typeof computedValue.value === 'string') {
@@ -64,22 +80,34 @@ const onClickCopy = () => {
 
 const scrollDown = () => {
   tapIf(contentRef.value, (el) => {
-    // 100px from bottom (temp)
-    if (el.clientHeight + el.scrollTop + 100 > el.scrollHeight) {
+    if (isAnchored.value) {
       el.scrollTo(el.scrollLeft, el.scrollHeight);
     }
   });
 };
 
-onMounted(scrollDown);
+watch(
+  computedValue,
+  () => {
+    requestAnimationFrame(() => {
+      scrollDown();
+    });
+  },
+  { immediate: true },
+);
 
-onUpdated(scrollDown);
+const onContentScroll = (ev: Event) => {
+  const el = ev.target as HTMLElement;
+  isAnchored.value = el.scrollTop + 20 /* ~ 1 line height */ >= el.scrollHeight - el.offsetHeight;
+};
 </script>
 
 <template>
-  <div class="pl-log-view" :class="{ 'has-error': computedError }">
-    <MaskIcon24 title="Copy content" class="pl-log-view__copy" name="clipboard" @click="onClickCopy" />
+  <div ref="root" class="pl-log-view" :class="{ 'has-error': computedError }">
+    <label v-if="label"> {{ label }} </label>
+    <DoubleContour class="pl-log-view__contour" />
+    <PlMaskIcon24 title="Copy content" class="pl-log-view__copy" name="clipboard" @click="onClickCopy" />
     <div v-if="computedError" class="pl-log-view__error">{{ computedError }}</div>
-    <div v-else ref="contentRef" class="pl-log-view__content">{{ computedValue }}</div>
+    <div v-else ref="contentRef" class="pl-log-view__content" @scroll="onContentScroll">{{ computedValue }}</div>
   </div>
 </template>