+++ etc/block-components/ui-examples/model/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @milaboratories/milaboratories.ui-examples.model
 
+## 1.0.28
+
+### Patch Changes
+
+- 4a745bf: Added password type for PlTextField
+
+## 1.0.27
+
+### Patch Changes
+
+- c16492e: Update icon-set
+
+## 1.0.26
+
+### Patch Changes
+
+- f341244: AgGrid: migrate to the Theming Api
+
 ## 1.0.25
 
 ### Patch Changes