+++ etc/blocks/enter-numbers/CHANGELOG.md
@@ -1,5 +1,73 @@
 # @milaboratories/milaboratories.test-enter-numbers
 
+## 1.0.64
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.64
+
+## 1.0.63
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.63
+
+## 1.0.62
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.62
+
+## 1.0.61
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.61
+
+## 1.0.60
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.60
+
+## 1.0.59
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.59
+
+## 1.0.58
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.58
+
+## 1.0.57
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.57
+
+## 1.0.56
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.56
+
+## 1.0.55
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.55
+
+## 1.0.54
+
+### Patch Changes
+
+- @milaboratories/milaboratories.test-enter-numbers.ui@1.0.54
+- @milaboratories/milaboratories.test-enter-numbers.workflow@1.0.3
+- @milaboratories/milaboratories.test-enter-numbers.model@1.0.15
+
 ## 1.0.53
 
 ### Patch Changes