+++ lib/node/pl-tree/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @milaboratories/pl-tree
 
+## 1.4.9
+
+### Patch Changes
+
+- Updated dependencies [0573960]
+  - @milaboratories/pl-client@2.5.8
+
+## 1.4.8
+
+### Patch Changes
+
+- Updated dependencies [75b1646]
+  - @milaboratories/ts-helpers@1.1.1
+  - @milaboratories/computable@2.2.1
+  - @milaboratories/pl-client@2.5.7
+
+## 1.4.7
+
+### Patch Changes
+
+- Updated dependencies [50f7459]
+  - @milaboratories/pl-client@2.5.6
+
 ## 1.4.6
 
 ### Patch Changes