+++ sdk/workflow-tengo/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @platforma-sdk/workflow-tengo
 
+## 1.7.4
+
+### Patch Changes
+
+- 554e5d3: PFrames version bump
+
+## 1.7.3
+
+### Patch Changes
+
+- 68de5a5: pfconv version bump
+
+## 1.7.2
+
+### Patch Changes
+
+- 50f7459: Dep upgrade, vitest in particular
+
 ## 1.7.1
 
 ### Patch Changes