+++ tools/block-tools/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @platforma-sdk/block-tools
 
+## 2.3.21
+
+### Patch Changes
+
+- Updated dependencies [75b1646]
+  - @milaboratories/ts-helpers@1.1.1
+  - @milaboratories/ts-helpers-oclif@1.1.10
+
+## 2.3.20
+
+### Patch Changes
+
+- 50f7459: Dep upgrade, vitest in particular
+
 ## 2.3.19
 
 ### Patch Changes