+++ tools/pl-bootstrap/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @platforma-sdk/bootstrap
 
+## 2.5.10
+
+### Patch Changes
+
+- Updated dependencies [5a4d01b]
+  - @milaboratories/pl-local@1.5.2
+
+## 2.5.9
+
+### Patch Changes
+
+- @milaboratories/pl-local@1.5.1
+
+## 2.5.8
+
+### Patch Changes
+
+- 50f7459: Dep upgrade, vitest in particular
+
 ## 2.5.7
 
 ### Patch Changes